Final Project - Drug Search
Goal of project was to create a simple GUI that would allow people to enter in symptoms and then the search engine would return a list of drugs that could be used to treat those symptoms.  Search is limited to the total number of drugs stored, which currently is about 200 I think.  I tried uploading the drug files also but github said it was too many files.

Project does use Maven, so some set up may be required.